card_of_vologda/requirements.txt
Django==3.2.13
Pillow==9.1.0
future~=0.18.2
pip~=21.3.1
wheel~=0.37.1
lxml~=4.7.1
keyring~=23.5.0
pytz~=2022.1
pkginfo~=1.8.2
sqlparse~=0.4.2
psycopg2~=2.9.3
asgiref~=3.5.0
colorama~=0.4.4
PyYAML~=6.0
Jinja2~=3.0.3
serializers~=0.2.4
docutils~=0.18.1
setuptools~=49.2.1
django-softdelete~=0.9.2